#16717 defer provider instatiation
